\name{CaretModel}
\alias{CaretModel}
\title{CaretModel}
\description{
  A class in the PredictiveModel class hierarchy that wraps
  models returned from caret. This class allows all methods
  implemented in the caret machine learning package to plug
  into pipelines using the PredictiveModeling API by
  wrapping all caret methods to implement customTrain() and
  customPredict() methods.
}
\author{
  Adam Margolin
}

