\name{defaultTrainControl}
\alias{defaultTrainControl}
\title{Create a default trControl object for use with caret}
\usage{
  defaultTrainControl()
}
\value{
  trainControl
}
\description{
  Create a default trControl object for use with caret
}
\author{
  Adam Margolin
}
\seealso{
  caret::trainControl
}

