\name{exactConcordanceIndex}
\alias{exactConcordanceIndex}
\title{Compute the exact value of concordance index}
\usage{
  exactConcordanceIndex(predictions, observations)
}
\arguments{
  \item{predictions}{a vector of predictions}

  \item{observations}{a matrix containing time and status
  columns}
}
\description{
  Compute the exact value of concordance index
}
\author{
  Ben Sauerwine
}

