\name{filterNasFromMatrix}
\alias{filterNasFromMatrix}
\title{removes rows and or columns from a matrix that contain NA values}
\usage{
  filterNasFromMatrix(dataMatrix, filterBy = "rows")
}
\description{
  removes rows and or columns from a matrix that contain NA
  values
}
\author{
  Adam Margolin
}

